preliminary pub-sub approach -- uses a bridge-event implementation; refer to 04228a52ed1c49f8c1e1d5da38ece93ae98a8e68 for the core-integrated emit system
