+++ packages/devtools/src/index.ts
@@ -1,0 +1,2 @@
+export { initializeDevtools } from './devtools.js';
+
