+++ packages/devtools/tsconfig.json
@@ -1,0 +1,10 @@
+{
+  "extends": "../../tsconfig.base.json",
+  "include": ["src/**/*"],
+  "compilerOptions": {
+    "outDir": "dist",
+    "strict": true
+  },
+  "references": [{ "path": "../mosaic/core" }]
+}
+
